+++ src/index.ts
@@ -30,4 +30,4 @@ export * as waku_relay from "./lib/waku_relay";
 export { WakuRelay } from "./lib/waku_relay";
 
 export * as waku_store from "./lib/waku_store";
-export { PageDirection, WakuStore, StoreCodecs } from "./lib/waku_store";
+export { PageDirection, WakuStore, StoreCodec } from "./lib/waku_store";